+++ lib/leap_web/components/main/post/state.ex
@@ -1,19 +1,52 @@
-defmodule LeapWeb.Components.Main.Post.Mutation do
+defmodule LeapWeb.Components.Main.Post.State do
   @moduledoc """
   Mutations for Post
-  Need to make distinction between mutations that update the changeset or not
   """
 
-  alias LeapWeb.Components.Main.Post.State
+  use TypedStruct
+
+  alias __MODULE__
+  alias LeapWeb.Components.State, as: StateBehaviour
 
   alias Leap.Content
   alias Leap.Content.Schema.Post
   alias Leap.Group
   alias Leap.Group.Schema.Category
+  alias Leap.Accounts.Schema.User
+
+  @behaviour StateBehaviour
+
+  @typedoc "Post state"
+  typedstruct do
+    field :id, String.t(), enforce: true
+    field :module, module(), enforce: true
+    field :current_user, User.t()
+    field :post, Post.t(), enforce: true
+    field :post_changeset, Ecto.Changeset.t(Post.t())
+    field :post_behaviour, atom(), default: :show_post, enforce: true
+    field :categories, [Category.t()], default: []
+  end
 
   defguard is_present(value) when is_binary(value) and bit_size(value) > 0
 
-  @spec commit(atom(), any(), State.t()) :: State.t()
+  @impl StateBehaviour
+  def init(%{id: id, post: post, current_user: current_user}) do
+    changeset = Content.change_post(post)
+
+    state = %State{
+      id: id,
+      module: LeapWeb.Components.Main.Post,
+      current_user: current_user,
+      post: post,
+      post_changeset: changeset,
+      post_behaviour: post_behaviour(post),
+      categories: Group.all(Category)
+    }
+
+    {:ok, state}
+  end
+
+  @impl StateBehaviour
   def commit(:update_post, post_params, state) do
     case Content.update_post(state.current_user, state.post, post_params) do
       {:ok, post} ->
@@ -66,7 +99,7 @@ defmodule LeapWeb.Components.Main.Post.Mutation do
 
       {:error, %Ecto.Changeset{} = changeset} ->
         state = %State{state | post_changeset: changeset}
-        {:ok, state}
+        {:ok, {state, {:danger, "Error publishing post"}}}
 
       {:error, message} ->
         {:error, message}
@@ -100,4 +133,8 @@ defmodule LeapWeb.Components.Main.Post.Mutation do
   defp with_preloads(%Post{} = post) do
     Content.with_preloads(post, [:category], force: true)
   end
+
+  defp post_behaviour(%Post{state: :new}), do: :edit_post
+  defp post_behaviour(%Post{state: :draft}), do: :edit_post
+  defp post_behaviour(%Post{state: :published}), do: :show_post
 end
